+++ pkgs/dart_tooling_mcp_server/lib/src/server.dart
@@ -27,8 +27,8 @@ final class DartToolingMCPServer extends MCPServer
         PubSupport,
         DartToolingDaemonSupport
     implements ProcessManagerSupport {
-  DartToolingMCPServer({
-    required super.channel,
+  DartToolingMCPServer(
+    super.channel, {
     @visibleForTesting this.processManager = const LocalProcessManager(),
   }) : super.fromStreamChannel(
          implementation: ServerImplementation(
@@ -43,7 +43,7 @@ final class DartToolingMCPServer extends MCPServer
   static Future<DartToolingMCPServer> connect(
     StreamChannel<String> mcpChannel,
   ) async {
-    return DartToolingMCPServer(channel: mcpChannel);
+    return DartToolingMCPServer(mcpChannel);
   }
 
   @override