+++ pkgs/dart_tooling_mcp_server/test/test_harness.dart
@@ -390,7 +390,7 @@ Future<ServerConnectionPair> _initializeMCPServer(
       serverController.sink,
     );
     server = DartToolingMCPServer(
-      channel: serverChannel,
+      serverChannel,
       processManager: TestProcessManager(),
     );
     addTearDown(server.shutdown);