+++ pkgs/dart_mcp/CHANGELOG.md
@@ -18,6 +18,9 @@
   `ResourceListChangedNotification`s and `ResourceUpdatedNotification`s. The
   delay can be modified by overriding
   `ResourcesSupport.resourceUpdateThrottleDelay`.
+- Add `Sink<String> protocolLogSink` parameters to server constructor and client
+  connection methods, which can be used to capture protocol messages for
+  debugging purposes.
 - Only send notifications if the peer is still connected. Fixes issues where
   notifications are delayed due to throttling and the client has since closed.
 - **Breaking**: Fixed paginated result subtypes to use `nextCursor` instead of
@@ -29,8 +32,14 @@
   which has all supported versions and whether or not they are supported.
 - **Breaking**: Change `InitializeRequest` and `InitializeResult` to take a
   `ProtocolVersion` instead of a string.
-- **Breaking**: Change the `InitializaeResult`'s `instructions` to `String?` to reflect
+- **Breaking**: Change the `InitializeResult`'s `instructions` to `String?` to reflect
   that not all servers return instructions.
+- **Breaking**: Change `MCPBase` to accept a `StreamChannel<String>` instead of
+  a `Peer`, and construct its own `Peer`.
+- **Breaking**: Add `protocolLogSink` optional parameter to connect methods on
+  `MCPClient`.
+- **Breaking**: Move `channel` parameter on `MCPServer.new` to a positional
+  parameter for consistency.
 
 ## 0.1.0